OutlookAgent/requirements.txt
openai==1.109.1
python-dotenv==1.2.1
schedule==1.2.2
requests==2.32.5
Flask==3.1.3
